feat: :sparkles: Config: Configmaps - File injection with Volumes
ConfigMaps can be used with volumes to inject configuration data as files into containers. In this example, a ConfigMap named configmap-app-injection is mounted as a volume at the /family directory in the container. The ConfigMap's keys, members and ages, are both mapped to a file named family.txt within the /family directory, allowing the container to read the configuration data from the file.