feat: :sparkles: ServiceAccount: Creating the "server" namespace Svc Account
A ServiceAccount in Kubernetes is an identity used by Pods to authenticate with the Kubernetes API server and access cluster resources. Each Pod has a ServiceAccount associated with it, and it uses the ServiceAccount's credentials to interact with the API server.  A ServiceAccount is a Kubernetes object that defines a set of permissions and access to resources within a specific namespace. It is created within a namespace and can only be used within that namespace, unless it is explicitly granted permission to access resources in other namespaces or across the entire cluster.  ServiceAccounts are used to grant permissions to Pods and other resources in the cluster. By creating a ServiceAccount and granting it the appropriate permissions, you can control which resources can be accessed by Pods associated with that ServiceAccount. This helps to ensure that Pods can only access the resources they need to perform their intended functions, and nothing more.